# studyindexer/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StudyIndexer API",
    "description": "Document Processing and Semantic Search Service\n\nFeatures:\n* Document upload and processing (PDF, DOCX, TEXT, MARKDOWN)\n* Semantic search across documents\n* Course-specific document collections\n* Role-based access control\n* Document metadata management",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "DocumentStatus": {
        "type": "string",
        "enum": ["pending", "processing", "completed", "failed"]
      },
      "DocumentType": {
        "type": "string",
        "enum": ["pdf", "text", "docx", "markdown"]
      },
      "CollectionType": {
        "type": "string",
        "enum": ["general", "course", "personal", "faq"]
      },
      "BaseResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "default": false
              },
              "error": {
                "$ref": "#/components/schemas/ErrorDetail"
              }
            }
          }
        ]
      },
      "DocumentMetadata": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "course_id": {
            "type": "string",
            "nullable": true
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "collection": {
            "$ref": "#/components/schemas/CollectionType"
          },
          "custom_metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "DocumentInfo": {
        "type": "object",
        "properties": {
          "document_id": {
            "type": "string",
            "description": "Unique document identifier"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "metadata": {
            "$ref": "#/components/schemas/DocumentMetadata"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "DocumentListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "documents": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DocumentInfo"
                }
              },
              "total": {
                "type": "integer",
                "description": "Total number of documents"
              }
            }
          }
        ]
      },
...
      "StorageStats": {
        "type": "object",
        "properties": {
          "uploads": {
            "type": "string",
            "description": "Upload directory size"
          },
          "vectors": {
            "type": "string",
            "description": "Vector database size"
          },
          "total": {
            "type": "string",
            "description": "Total storage used"
          },
          "raw_total_bytes": {
            "type": "integer",
            "description": "Total storage in bytes"
          }
        }
      },
      "ProcessingStats": {
        "type": "object",
        "properties": {
          "pending": {
            "type": "integer",
            "default": 0,
            "description": "Number of pending documents"
          },
          "processing": {
            "type": "integer",
            "default": 0,
            "description": "Number of documents being processed"
          },
          "completed": {
            "type": "integer",
            "default": 0,
            "description": "Number of completed documents"
          },
          "failed": {
            "type": "integer",
            "default": 0,
            "description": "Number of failed documents"
          },
          "total": {
            "type": "integer",
            "description": "Total number of documents"
          }
        }
      },
      "CollectionStats": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Collection name"
          },
          "document_count": {
            "type": "integer",
            "description": "Number of documents"
          },
          "metadata": {
            "type": "object",
            "nullable": true,
            "description": "Collection metadata"
          }
        }
      },
      "SystemStats": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "total_documents": {
                "type": "integer",
                "description": "Total number of documents"
              },
              "total_collections": {
                "type": "integer",
                "description": "Number of collections"
              },
              "collections": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CollectionStats"
                },
                "description": "Collection statistics"
              },
              "storage_used": {
                "$ref": "#/components/schemas/StorageStats",
                "description": "Storage usage statistics"
              },
              "processing_status": {
                "$ref": "#/components/schemas/ProcessingStats",
                "description": "Processing statistics"
              },
              "embedding_model": {
                "type": "string",
                "description": "Current embedding model"
              },
              "embedding_device": {
                "type": "string",
                "description": "Current embedding device"
              },
              "last_update": {
                "type": "string",
                "format": "date-time",
                "description": "Last update timestamp"
              }
            }
          }
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/admin/documents": {
      "get": {
        "summary": "List all documents",
        "description": "Returns a list of all documents in the system (admin only)",
        "tags": ["admin"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of documents to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of documents to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/stats": {
      "get": {
        "summary": "Get system statistics",
        "description": "Returns system-wide statistics (admin only)",
        "tags": ["admin"],
        "responses": {
          "200": {
            "description": "System statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStats"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
} 